fix delete project confirmation

+++ resources/assets/js/components/projects/AllProjects.vue
@@ -142,11 +142,25 @@
             // location = "/project-hq/" + row.id;
         },
         destroy: function(row) {
-            axios.delete('/projects/' + row.id + '/delete')
+            var self = this;
+                swal({
+              title: 'Are you sure?',
+              text: "You won't be able to revert this!",
+              type: 'warning',
+              showCancelButton: true,
+              confirmButtonColor: '#3085d6',
+              cancelButtonColor: '#d33',
+              confirmButtonText: 'Yes, delete it!'
+            }).then(function (result) {
+              if (result) {
+                axios.delete('/projects/' + row.id + '/delete')
                 .then(response => {
                     swal('Success!', 'Project is Deleted!', 'success');
-                    this.getAllProjects();
+                    self.getAllProjects();
                 });
+              }
+          })
+            
         },
         cellClick: function(row, col) {
             var a = col.id;
